backport: dfl-cxl-cache.c
Substituting account_locked_vm() with afu_dma_adjust_locked_vm()
taking afu_dma_adjust_locked_vm() from the backport of dfl-afu-dma_region.c
