+++ drivers/fpga/dfl-afu-dma-region.c
@@ -17,6 +17,7 @@
 #include <linux/version.h>
 
 #include "dfl-afu.h"
+#include "dfl-dma-cxl-common.h"
 
 void afu_dma_region_init(struct dfl_feature_dev_data *fdata)
 {
@@ -26,44 +27,6 @@ void afu_dma_region_init(struct dfl_feature_dev_data *fdata)
 }
 
 #if LINUX_VERSION_CODE < KERNEL_VERSION(5, 3, 0) && RHEL_RELEASE_CODE < 0x803
-static long afu_dma_adjust_locked_vm(struct device *dev, long npages, bool incr)
-{
-	unsigned long locked, lock_limit;
-	int ret = 0;
-
-	/* the task is exiting. */
-	if (!current->mm)
-		return 0;
-
-	down_write(&current->mm->mmap_sem);
-
-	if (incr) {
-		locked = current->mm->locked_vm + npages;
-		lock_limit = rlimit(RLIMIT_MEMLOCK) >> PAGE_SHIFT;
-
-		if (locked > lock_limit && !capable(CAP_IPC_LOCK))
-			ret = -ENOMEM;
-		else
-			current->mm->locked_vm += npages;
-	} else {
-
-		if (WARN_ON_ONCE(npages > current->mm->locked_vm))
-			npages = current->mm->locked_vm;
-		current->mm->locked_vm -= npages;
-	}
-
-	dev_dbg(dev, "[%d] RLIMIT_MEMLOCK %c%ld %ld/%ld%s\n", current->pid,
-				incr ? '+' : '-',
-				npages << PAGE_SHIFT,
-				current->mm->locked_vm << PAGE_SHIFT,
-				rlimit(RLIMIT_MEMLOCK),
-				ret ? "- execeeded" : "");
-
-	up_write(&current->mm->mmap_sem);
-
-	return ret;
-}
-
 static int afu_dma_pin_pages(struct dfl_feature_dev_data *fdata,
 			     struct dfl_afu_dma_region *region)
 {
